Move repository url to correct location
I think some time in the past this was moved but this field was never
updated. Today we're getting npm errors trying to install things that
depend on this. I don't know if this is the fix but I do know this
change is more correct so I'm commiting it.